requirements.txt
fastapi~=0.104.1
sqlalchemy~=2.0.23
pymysql~=1.1.0
python-dotenv~=1.0.0
Flask-SQLAlchemy~=3.1.1
uvicorn~=0.24.0.post1
python-multipart~=0.0.6
passlib~=1.7.4
python-jose~=3.3.0
bcrypt~=4.1.1
influxdb-client~=1.39.0
pandas~=2.1.4
cryptography~=41.0.7
starlette~=0.27.0
pydantic~=2.5.1